type fix
Former-commit-id: d383ee46116fc6591328974e934efe83738cf925
Former-commit-id: 97f9c97275f9dcba1ced71298f59c6c828af61a9
Former-commit-id: f95a058a5de9cf92809b3f43c56aad92b5d5c198

+++ src/Types/KWDefs.jl
@@ -15,9 +15,9 @@
 # x     ::Array{Float64,1}          = Array{Float64,1}(undef, 0)
 
 struct SLDefs
-  port::Int
-  gap::Int
-  kai::Int
+  port::Int64
+  gap::Int64
+  kai::Int64
   mode::String
   refresh::Real
   x_on_err::Bool
@@ -28,15 +28,15 @@ struct KWDefs
   evw::Array{Float64,1}
   fmt::String
   mag::Array{Float64,1}
-  nev::Int
+  nev::Int64
   opts::String
   pha::String
   q::Char
   reg::Array{Float64,1}
   si::Bool
   src::String
-  to::Int
-  v::Int
+  to::Int64
+  v::Int64
   w::Bool
   y::Bool
 end
